requirements.txt
aioflask==0.4.0
asgiref==3.6.0
asn1crypto==1.5.1
asyncpg==0.27.0
bcrypt==4.0.1
cachelib==0.10.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
cloud-sql-python-connector
click==8.1.3
dependency-injector==4.39.1
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-Executor==1.0.0
Flask-GoogleStorage==0.1.2
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SQLAlchemy==3.0.2
Flask-SSLify==0.1.5
Flask-WTF==1.1.1
gevent==22.10.2
pg8000
google-api-core==2.11.0
google-auth==2.16.2
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
greenlet==2.0.2
greenletio==0.9.0
gunicorn
h11==0.14.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.24.2
pandas==1.5.0
protobuf==4.22.1
psycopg2-binary
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
rsa==4.9
scramp==1.4.4
six==1.16.0
SQLAlchemy==1.4.42
tenacity==8.2.2
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.10.0
zope.event==4.6
zope.interface==6.0
validate-email==1.3
